avoid warning when symlinking into symlink

+++ scripts/testing/setup-framework.sh
@@ -21,7 +21,7 @@ pushd $TEST_FRAMEWORK
 
 # scripts to run experiments
 # for now: (TODO: move to github)
-ln -sf /fs/lofn0/dmadl/mmt/bricks bricks
+[ ! -e bricks ] && ln -sf /fs/lofn0/dmadl/mmt/bricks bricks
 
 # experiment model staging area
 mkdir -p /fs/crom0/mmt/testing/models
